requirements.txt
flask==3.1.0
anthropic==0.43.0
gunicorn==23.0.0
PyMuPDF==1.25.3
python-docx==1.1.2
